project_cli/pyproject.toml
[project]
name = "cli"
version = "0.1.0"
description = ""
authors = [ { name = "Vasco Reynolds Brandao", email = "v.m.reynolds.brandao@student.rug.nl" } ]
dependencies = [ "click", "requests", "python-dotenv", "pandas" ]

[project.scripts]
mycli = "project_cli.main:main"